javadoc, neteja i millorar definició i tractament excepcions #80

+++ projecteorigen/projectebase-ejb/src/main/java/es/caib/projectebase/ejb/facade/UnitatOrganicaServiceFacadeBean.java
@@ -2,10 +2,14 @@ package es.caib.projectebase.ejb.facade;
 
 import es.caib.projectebase.commons.utils.Constants;
 import es.caib.projectebase.ejb.converter.UnitatOrganicaConverter;
+import es.caib.projectebase.ejb.interceptor.ExceptionTranslate;
 import es.caib.projectebase.ejb.interceptor.Logged;
+import es.caib.projectebase.ejb.repository.ProcedimentRepository;
 import es.caib.projectebase.ejb.repository.UnitatOrganicaRepository;
 import es.caib.projectebase.persistence.model.UnitatOrganica;
-import es.caib.projectebase.service.exception.UnitatOrganicaDuplicadaException;
+import es.caib.projectebase.service.exception.RecursNoTrobatException;
+import es.caib.projectebase.service.exception.UnitatDuplicadaException;
+import es.caib.projectebase.service.exception.UnitatTeProcedimentsException;
 import es.caib.projectebase.service.facade.UnitatOrganicaServiceFacade;
 import es.caib.projectebase.service.model.Ordre;
 import es.caib.projectebase.service.model.Pagina;
@@ -24,10 +28,12 @@ import java.util.Optional;
 /**
  * Implementació dels casos d'ús de manteniment de Unitats Orgàniques.
  * És responsabilitat d'aquesta capa definir el limit de les transaccions i la seguretat.
- *
+ * Les excepcions específiques es llancen mitjançant l'{@link ExceptionTranslate} que transforma
+ * els errors JPA amb les excepcions de servei com la {@link RecursNoTrobatException}
  * @author areus
  */
 @Logged
+@ExceptionTranslate
 @Stateless @Local(UnitatOrganicaServiceFacade.class)
 @TransactionAttribute(TransactionAttributeType.REQUIRES_NEW)
 @RolesAllowed(Constants.PBS_ADMIN)
@@ -37,12 +43,15 @@ public class UnitatOrganicaServiceFacadeBean implements UnitatOrganicaServiceFac
     private UnitatOrganicaRepository repository;
 
     @Inject
+    private ProcedimentRepository procedimentRepository;
+
+    @Inject
     private UnitatOrganicaConverter converter;
 
     @Override
-    public Long create(UnitatOrganicaDTO dto) throws UnitatOrganicaDuplicadaException {
+    public Long create(UnitatOrganicaDTO dto) throws UnitatDuplicadaException {
         if (repository.findByCodiDir3(dto.getCodiDir3()).isPresent()) {
-            throw new UnitatOrganicaDuplicadaException(dto.getCodiDir3());
+            throw new UnitatDuplicadaException(dto.getCodiDir3());
         }
 
         UnitatOrganica unitat = converter.toEntity(dto);
@@ -51,13 +60,16 @@ public class UnitatOrganicaServiceFacadeBean implements UnitatOrganicaServiceFac
     }
 
     @Override
-    public void update(UnitatOrganicaDTO dto) {
+    public void update(UnitatOrganicaDTO dto) throws RecursNoTrobatException {
         UnitatOrganica unitat = repository.getReference(dto.getId());
         converter.updateFromDTO(unitat, dto);
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Long id) throws UnitatTeProcedimentsException, RecursNoTrobatException {
+        if (procedimentRepository.countByUnitat(id) > 0L) {
+            throw new UnitatTeProcedimentsException(id);
+        }
         UnitatOrganica unitat = repository.getReference(id);
         repository.delete(unitat);
     }
